real-time database configured, empty input submit bug fixed

+++ src/components/TodoList.js
@@ -3,7 +3,8 @@ import { useSelector, useDispatch } from 'react-redux'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faRemove, faCheck } from '@fortawesome/free-solid-svg-icons'
 import { removeTodo, completedTodo, getTodos } from '../redux/todoListSlice'
-import { todoListRef } from "../firebase"
+import { db } from "../firebase"
+import { collection, onSnapshot, setDoc, deleteDoc, doc } from "firebase/firestore"
 
 
 function TodoList() {
@@ -12,46 +13,55 @@ function TodoList() {
   const removeIcon = <FontAwesomeIcon icon={faRemove} />
   const Check = <FontAwesomeIcon icon={faCheck} />
 
-  const updateTodos = () => {
-    todoListRef.once('value', (snapshot) => {
-      dispatch(getTodos(snapshot.val()))
-    }, (errorObject) => {
-      console.log('The read failed: ' + errorObject.name);
-    });
-  }
   useEffect(() => {
-    updateTodos()
+    const databaseCollection = collection(db, "todo-list")
+    //onSnapshot function watch changes on database collection
+    onSnapshot(databaseCollection, (snapshot) => {
+      snapshot.docChanges().forEach(change => {
+        const id = change.doc.id
+        if (change.type === "added") {
+          const todo = { ...change.doc.data(), id: id }
+          dispatch(getTodos(todo))
+        } else if (change.type === "modified") {
+          dispatch(completedTodo(id))
+        } else if (change.type === "removed") {
+          dispatch(removeTodo(id))
+        }
+      })
+    });
   }, [])
 
-  const handleRemove = (id) => {
-    dispatch(removeTodo(id))
-    updateTodos()
+  const handleRemove = async (id) => {
+    const docRef = doc(db, "todo-list", id)
+    await deleteDoc(docRef)
   }
-  const handleComplete = (id, completed) => {
-    dispatch(completedTodo({ id: id, completed: completed }))
-    updateTodos()
+
+  const handleComplete = async (todo) => {
+    const docRef = doc(db, "todo-list", todo.id)
+    const payload = { text: todo.text, completed: !todo.completed }
+    await setDoc(docRef, payload)
   }
 
   return (
     <div>
       <ul className="lists list-group todos">
         {
-          todoList ?
-            Object.keys(todoList).map(id =>
+          todoList.length ?
+            todoList.map(todo =>
               < li
-                key={id}
-                className={`todo list-group-item d-flex justify-content-between align-items-center ${todoList[id].completed ? 'completed' : ''}`}>
+                key={todo.id}
+                className={`todo list-group-item d-flex justify-content-between align-items-center ${todo.completed ? 'completed' : ''}`}>
 
-                <span>{todoList[id].text}</span>
+                <span>{todo.text}</span>
                 <div className='icons'>
                   <div
-                    className={`check icon ${todoList[id].completed ? 'completed' : ''}`}
-                    onClick={() => handleComplete(id, todoList[id].completed)}
+                    className={`check icon ${todo.completed ? 'completed' : ''}`}
+                    onClick={() => handleComplete(todo)}
                   >
                     {Check}
                   </div>
                   <div
-                    onClick={() => handleRemove(id)}
+                    onClick={() => handleRemove(todo.id)}
                     className='remove icon'
                   >
                     {removeIcon}
